+++ frontend/src/app/modules/aspect-configuration/services/aspect-configuration.service.spec.ts
@@ -5,13 +5,29 @@ import {HttpClientTestingModule, HttpTestingController} from "@angular/common/ht
 import {Application} from "../../../models/application.model";
 import {HttpRequest} from "@angular/common/http";
 import {BrowserInfoDto} from "../../../models/browser.model";
+import {ApplicationService} from "../../../services/application.service";
+import {
+  ExtendedPerformanceAspect,
+  PerformanceAspect,
+  PerformanceAspectType
+} from "../../../models/perfomance-aspect.model";
+import {BehaviorSubject, ReplaySubject} from "rxjs";
+import {Page} from "../../../models/page.model";
 
 describe('AspectConfigurationService', () => {
 
   beforeEach(() => {
     TestBed.configureTestingModule({
       providers: [
-        AspectConfigurationService
+        AspectConfigurationService,
+        {
+          provide: ApplicationService,
+          useValue: {
+            selectedApplication$: new ReplaySubject<Application>(1),
+            selectedPage$: new ReplaySubject<Page>(1),
+            performanceAspectsForPage$: new BehaviorSubject<PerformanceAspect[]>([])
+          }
+        }
       ],
       imports: [
         HttpClientTestingModule
@@ -25,19 +41,21 @@ describe('AspectConfigurationService', () => {
   });
 
   it('should be able to load application by id', inject(
-    [HttpTestingController, AspectConfigurationService],
-    (httpMock: HttpTestingController, service: AspectConfigurationService) => {
+    [HttpTestingController, AspectConfigurationService, ApplicationService],
+    (httpMock: HttpTestingController, service: AspectConfigurationService, applicationService: ApplicationService) => {
 
       const appId = 1;
       const appName = "test-app";
 
-      service.loadApplication("1").subscribe((app: Application) => {
+      applicationService.selectedApplication$.subscribe((app: Application) => {
         expect(app.id).toBe(appId);
         expect(app.name).toBe(appName);
         expect(app.pageCount).toBeNull();
         expect(app.dateOfLastResults).toBeNull();
         expect(app.csiConfigurationId).toBeNull();
       });
+
+      service.loadApplication("1");
       const req = httpMock.expectOne((req: HttpRequest<any>) => {
         expect(req.url).toBe("/applicationDashboard/rest/getApplication");
         return true;
@@ -49,8 +67,8 @@ describe('AspectConfigurationService', () => {
   ));
 
   it('should be able to load application by id and extracting all information from dto', inject(
-    [HttpTestingController, AspectConfigurationService],
-    (httpMock: HttpTestingController, service: AspectConfigurationService) => {
+    [HttpTestingController, AspectConfigurationService, ApplicationService],
+    (httpMock: HttpTestingController, service: AspectConfigurationService, applicationService: ApplicationService) => {
 
       const appId = 1;
       const appName = "test-app";
@@ -58,13 +76,16 @@ describe('AspectConfigurationService', () => {
       const pageCount = 5;
       const csiConfId = 53;
 
-      service.loadApplication("1").subscribe((app: Application) => {
+      applicationService.selectedApplication$.subscribe((app: Application) => {
         expect(app.id).toBe(appId);
         expect(app.name).toBe(appName);
         expect(app.pageCount).toBe(pageCount);
         expect(app.dateOfLastResults).toEqual(expectedDate);
         expect(app.csiConfigurationId).toBe(csiConfId)
       });
+
+      service.loadApplication("1");
+
       const req = httpMock.expectOne((req: HttpRequest<any>) => {
         expect(req.url).toBe("/applicationDashboard/rest/getApplication");
         return true;
@@ -101,13 +122,19 @@ describe('AspectConfigurationService', () => {
       ];
 
       service.loadBrowserInfos();
+      // const requests: TestRequest[] = httpMock.match("/aspectConfiguration/rest/getBrowserInformations");
       const req = httpMock.expectOne((req: HttpRequest<any>) => {
+        console.log('req.url=' + req.urlWithParams);
         expect(req.url).toBe("/aspectConfiguration/rest/getBrowserInformations");
         return true;
       });
+      // requests.forEach((req: TestRequest) => {
+      // });
       expect(req.request.method).toEqual('GET');
       req.flush(browserInfos);
+
       httpMock.verify();
+
       const infosFromObservable: BrowserInfoDto[] = service.browserInfos$.getValue();
       expect(infosFromObservable.length).toBe(2);
       const info1: BrowserInfoDto = infosFromObservable.filter((info: BrowserInfoDto) => info.browserId == 1)[0];
@@ -119,6 +146,141 @@ describe('AspectConfigurationService', () => {
       expect(info2.operatingSystem).toBe("Android");
       expect(info2.deviceType).toEqual({name: 'Galaxy Tab A', icon: 'tablet'});
     }))
+
+  it('should provide unique AspectTypes correctly according to given PerformanceAspects', inject(
+    [AspectConfigurationService],
+    (service: AspectConfigurationService) => {
+      const type1: PerformanceAspectType = {name: 'PAGE_CONSTRUCTION_STARTED', icon: 'hourglass'};
+      const type2: PerformanceAspectType = {name: 'PAGE_SHOWS_USEFUL_CONTENT', icon: 'hourglass'};
+      const aspectsOfTwoDifferentTypes = [
+        {
+          id: 1,
+          pageId: 1,
+          jobGroupId: 1,
+          browserId: 1,
+          measurand: {id: 'DOC_COMPLETE', name: 'Document complete'},
+          performanceAspectType: type1,
+          persistent: true,
+          browserName: 'browser1',
+          operatingSystem: 'Android',
+          deviceType: {name: 'Smartphone', icon: 'mobile'}
+        },
+        {
+          id: 1,
+          pageId: 2,
+          jobGroupId: 1,
+          browserId: 1,
+          measurand: {id: 'DOC_COMPLETE', name: 'Document complete'},
+          performanceAspectType: type1,
+          persistent: true,
+          browserName: 'browser1',
+          operatingSystem: 'Android',
+          deviceType: {name: 'Smartphone', icon: 'mobile'}
+        },
+        {
+          id: 1,
+          pageId: 3,
+          jobGroupId: 1,
+          browserId: 1,
+          measurand: {id: 'DOC_COMPLETE', name: 'Document complete'},
+          performanceAspectType: type1,
+          persistent: true,
+          browserName: 'browser1',
+          operatingSystem: 'Android',
+          deviceType: {name: 'Smartphone', icon: 'mobile'}
+        },
+        {
+          id: 1,
+          pageId: 4,
+          jobGroupId: 1,
+          browserId: 1,
+          measurand: {id: 'DOC_COMPLETE', name: 'Document complete'},
+          performanceAspectType: type2,
+          persistent: true,
+          browserName: 'browser1',
+          operatingSystem: 'Android',
+          deviceType: {name: 'Smartphone', icon: 'mobile'}
+        }];
+      service.performanceAspects$.next(aspectsOfTwoDifferentTypes);
+      service.initAspectTypes();
+      service.uniqueAspectTypes$.subscribe((aspectTypes: PerformanceAspectType[]) => {
+        expect(aspectTypes.length).toBe(2);
+        expect(aspectTypes.filter((type: PerformanceAspectType) => type == type1).length).toBe(1);
+        expect(aspectTypes.filter((type: PerformanceAspectType) => type == type2).length).toBe(1);
+      })
+    }));
+
+  describe('aspect extension', () => {
+    it('should not provide any extended aspects if just BrowserInfos and no PerformanceAspects exist', inject(
+      [AspectConfigurationService],
+      (service: AspectConfigurationService) => {
+        const browserInfos: BrowserInfoDto[] = [{
+          browserId: 1,
+          browserName: 'Chrome',
+          operatingSystem: 'Windows',
+          deviceType: {name: 'Desktop', icon: 'desktop'}
+        }];
+        const extendedAspects: ExtendedPerformanceAspect[] = service.extendAspects([], browserInfos);
+        expect(extendedAspects.length).toBe(0)
+      }));
+    it('should not provide any extended aspects if just PerformanceAspects and no BrowserInfos exist', inject(
+      [AspectConfigurationService],
+      (service: AspectConfigurationService) => {
+        const aspects: PerformanceAspect[] = [{
+          id: 1,
+          pageId: 1,
+          jobGroupId: 1,
+          browserId: 1,
+          measurand: {id: 'DOC_COMPLETE', name: 'Document complete'},
+          performanceAspectType: {name: 'PAGE_CONSTRUCTION_STARTED', icon: 'hourglass'},
+          persistent: true,
+        }];
+        const extendedAspects: ExtendedPerformanceAspect[] = service.extendAspects(aspects, []);
+        expect(extendedAspects.length).toBe(0)
+      }));
+    it('should extend aspects correctly only if matching BrowserInfo exists', inject(
+      [AspectConfigurationService],
+      (service: AspectConfigurationService) => {
+        const idOfExtended: number = 1;
+        const idOfNotExtended: number = 2;
+        const aspects: PerformanceAspect[] = [
+          {
+            id: 1,
+            pageId: 1,
+            jobGroupId: 1,
+            browserId: idOfExtended,
+            measurand: {id: 'DOC_COMPLETE', name: 'Document complete'},
+            performanceAspectType: {name: 'PAGE_CONSTRUCTION_STARTED', icon: 'hourglass'},
+            persistent: true,
+          },
+          {
+            id: 1,
+            pageId: 1,
+            jobGroupId: 1,
+            browserId: idOfNotExtended,
+            measurand: {id: 'DOC_COMPLETE', name: 'Document complete'},
+            performanceAspectType: {name: 'PAGE_CONSTRUCTION_STARTED', icon: 'hourglass'},
+            persistent: true,
+          }
+        ];
+        const browserInfos: BrowserInfoDto[] = [{
+          browserId: idOfExtended,
+          browserName: 'Chrome',
+          operatingSystem: 'Windows',
+          deviceType: {name: 'Desktop', icon: 'desktop'}
+        }];
+        const extendedAspects: ExtendedPerformanceAspect[] = service.extendAspects(aspects, browserInfos);
+        expect(extendedAspects.length).toBe(2);
+        const extendedAspect: ExtendedPerformanceAspect = extendedAspects.filter((aspect: PerformanceAspect) => aspect.browserId == idOfExtended)[0];
+        expect(extendedAspect.browserName).toBe('Chrome');
+        expect(extendedAspect.operatingSystem).toBe('Windows');
+        expect(extendedAspect.deviceType).toEqual({name: 'Desktop', icon: 'desktop'});
+        const notExtendedAspect: ExtendedPerformanceAspect = extendedAspects.filter((aspect: PerformanceAspect) => aspect.browserId == idOfNotExtended)[0];
+        expect(notExtendedAspect.browserName).toBe('Unknown');
+        expect(notExtendedAspect.operatingSystem).toBe('Unknown');
+        expect(notExtendedAspect.deviceType).toEqual({name: 'Unknown', icon: 'question'});
+      }));
+  });
 
 })
 ;
